Improved message sending, includes size
Now the byte[] array that gets sent is prepended with the size of the
message. The threads that read messages will continue attempting to read
until that many bytes are read before trying to convert it back into a
TcpMessage. Hopefully this solves the issue of high-latency connections
having errors when receiving messages.